+++ src/main/java/net/sneezewipe/steelwitchesplus/datagen/ModModelProvider.java
@@ -8,6 +8,8 @@ import net.minecraft.data.client.ItemModelGenerator;
 import net.minecraft.data.client.Models;
 import net.minecraft.item.ArmorItem;
 import net.sneezewipe.steelwitchesplus.block.ModBlocks;
+import net.sneezewipe.steelwitchesplus.block.custom.ArtichokeCropBlock;
+import net.sneezewipe.steelwitchesplus.block.custom.GarlicCropBlock;
 import net.sneezewipe.steelwitchesplus.item.ModItems;
 
 public class ModModelProvider extends FabricModelProvider {
@@ -19,6 +21,9 @@ public class ModModelProvider extends FabricModelProvider {
     public void generateBlockStateModels(BlockStateModelGenerator blockStateModelGenerator) {
         BlockStateModelGenerator.BlockTexturePool quartzPillarPool = blockStateModelGenerator.registerCubeAllModelTexturePool(Blocks.QUARTZ_PILLAR);
         quartzPillarPool.wall(ModBlocks.QUARTZ_WALL);
+
+        blockStateModelGenerator.registerCrop(ModBlocks.ARTICHOKE_CROP, ArtichokeCropBlock.AGE, 0, 1, 2, 3, 4);
+        blockStateModelGenerator.registerCrop(ModBlocks.GARLIC_CROP, GarlicCropBlock.AGE, 0, 1, 2, 3, 4);
     }
 
     @Override
@@ -30,6 +35,7 @@ public class ModModelProvider extends FabricModelProvider {
         itemModelGenerator.register(ModItems.ATTUNED_STONE_DARK, Models.GENERATED);
         itemModelGenerator.register(ModItems.ATTUNED_STONE_LIGHT, Models.GENERATED);
         itemModelGenerator.register(ModItems.BAKED_CLAY_JAR, Models.GENERATED);
+        itemModelGenerator.register(ModItems.GARLIC, Models.GENERATED);
         itemModelGenerator.register(ModItems.CLAY_JAR, Models.GENERATED);
         itemModelGenerator.register(ModItems.GLASS_JAR, Models.GENERATED);
         itemModelGenerator.register(ModItems.SCULK_POWDER, Models.GENERATED);