+++ CHANGELOG.md
@@ -4,7 +4,6 @@
 * Upgraded mod to Minecraft v1.21.1
 * New items and blocks with crafting recipes:
   * Amethyst Dust
-  * Artichoke
   * Dark Attunement Stone
   * Light Attunement Stone
   * Baked Clay Jar
@@ -12,7 +11,8 @@
   * Glass Jar
   * Sculk Powder
   * Weeping Powder
-* Sadly, can't figure out how to render custom elytra, so those are gone for now :(
+* New crops and seeds:
+  * Artichoke
 
 ### v1.1.0
 As far as I'm concerned, v1.1.0 is the first release, since I don't know the differences between v1.1.0 and the actual first release v1.0.0 anymore lol.